patina-dxe-core: Add ARM64 64 KB Runtime Allocation Granularity (#552)
## Description

This patch adds support and unit tests for ARM64 64 KB runtime page
allocation granularity, which is required by the UEFI spec to ensure
that 16 KB and 64 KB OSes can boot.

Note, this granularity is not tested at all in edk2 and has been a
source of many bugs. This is testable in patina with the
with_granularity_modulation fn.

This patch also adds debug_asserts to places where the FSB fails;
otherwise we lose what the allocator was doing because Rust core module
code will receive the `AllocError` and then call the
`alloc_error_handler`, which panics, but we've lost the stack at that
point. We see the top level allocation that failed, but not what failed
within the allocator, so debugging is useless at that point. With the
debug_asserts, we can see what failed in the allocator and debug from
there. There is a concern that in production code we will panic on
`AllocError`s, but this seems to be a fundamental design of the core
Rust code, so I do not think Patina can do anything else.

This patch dropped some unit tests that weren't testing anything
meaningful and removed a couple of test asserts that cannot be tested
now due to debug_asserting in alloc errors.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested by booting Windows and Linux (4KB and 64KB for ARM64) on Q35 and
SBSA and viewing the EFI_MEMORY_MAP and MAT table to confirm all runtime
types have 4 KB alignment and length on x86 and 64 KB on ARM64.

## Integration Instructions

This closes #524 and so allows platforms to boot 16 KB and 64 KB
Linuxes.