Move content from `patina_performance` to `patina_sdk`
A lot of content in the `patina_performance` component is not solely
within the scope of the component. It makes understanding and
maintaining the component more difficult and places like
`patina_dxe_core` have dependencies on the component module to access
content that should be in a more broadly accessible area.

This change moves content out of `patina_performance` to resolve these
issues and support an upcoming refactor of the component.

Note: This also makes some additional clean up and changes such as
restructuring the namespace, increasing documentation, and some
minor code changees. However, this is not meant to resolve all of
the work that should be done to refactor the performance code. For
example, the code depends on atomic globals for maintaining some
state. An alternative solution would be cleaner. This change simply
collects those static globals into a `globals` module so they are
clearly visible as such.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>