patina_dxe_core: Enable Null Pointer Detection
This commit disallows allocating page 0 in order to leave it
unmapped to detect null pointers dereferencing. This also
ensured that when EFI_MEMORY_RP is applied via set_memory_space_attributes,
that Patina DXE Core will unmap the region, not simply attempt to apply
the attribute.

While changing the allocation code, this also changes the parameter name
of alignment -> align_shift, except in PI spec defined APIs, as that
better reflects the usage of that parameter.