Update arm-gic to 0.5.0 (#600)
## Description

Updates patina AARCH64 GIC support to use 0.5.0 of arm-gic crate. There
have been significant API changes since arm-gic crate was first
integrated. This PR brings patina up to the latest API. In the course of
integrating this, found and filed the following bug to upstream (thanks,
@kuqin12 ):
[https://github.com/google/arm-gic/issues/67](https://github.com/google/arm-gic/issues/67).
This issue is not fatal to operation but will cause error messages on
systems with GIC RAS features enabled.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Built and tested boot on an AARCH64 hardware platform. 

## Integration Instructions

N/A