Remove the exclusion of patina_sdk_macro from code coverage (#611)
## Description

The patina_sdk_macro crate keeps being added as an exclusion in the
Makefile because an older version of cargo tarpaulin would crash when
testing macro crates. When local developers hit this issue, they are
adding the crate to the excluded list, when the real solution is to
upgrade to the version of cargo tarpaulin specified in the
rust-toolchain.toml file. This change has snuck in and been merged a
time or two I think. So just removing the exclusion again.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI Passes

## Integration Instructions

N/A