patina_sdk: Do not call drop on failed PageAllocation type conversion
This commit prevents `drop` from being called on a `PageAllocation`
instance when it is consumed due to a failed conversion into a
different type (Happens when the byte length is insufficient). The drop
implementation of the `PageAllocation` type purposefully panics if it is
called, because it generally means someone made a PageAllocation, but
did not use it in any meaningful way.

However this handles the case where the `PageAllocation` is dropped, not
because it was never used, but because the conversion into another type
failed and the `PageAllocation` is dropped as a result (because it
consumes self).