patina_sdk: PageAllocation: Do not call `Drop` on unitialized memory (#622)
## Description

As referenced in #621, when filling the slice for `into_boxed_slice`,
the value being replaced (which is uninitialized) has it's `Drop`
implementation called because the function believes the data is
initialized. This commit updates the logic to use MaybeUninit, so that
when the value is inserted, the replaced, uninitialized data, does not
have it's `Drop` implementation called

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Failing test that detected this is now passing.

## Integration Instructions

N/A