patina_sdk: Improve pointer provenance
This commit enhances the pointer provenance tracking for allocations
made through the MemoryManager service in the Patina SDK. Pointer
provenance is crucial for ensuring memory safety, preventing undefined
behavior, allowing for enhanced compiler optimizations, and most
importantly, better evaluation of memory access patterns by static
analysis tools such as miri.

For more information on pointer provenance, see:
https://doc.rust-lang.org/std/ptr/index.html#provenance

With this commit, provenance is now exposed within the
`core_allocate_pages` function via the `expose_provenance` method, which
means any conversions of an address in the bounds of that allocation
will gain that pointer's provenance so long as it is converted via the
`with_exposed_provenance` or `with_exposed_provenance_mut` methods.
Simple conversions like `as *mut T` or `as *const T` should also pick up
the provenance, however it is better to use the provided methods to
ensure provenance is tracked correctly.

While this does help with provenance tracking once the allocation is
exposed to the caller of `core_allocate_pages`, further work should
eventually be done to either expose the pointer's provenance earlier in
the call-stack, or to generate and return a pointer type such as
`NonNull<T>` in the rust core, rather than a `usize`.