patina_sdk: Add size check to page allocation's into_raw_ptr
Other routines in the Memory Manager API will return `None` in the event that the
type of the pointer does not fit into the allocated pages. This moves that logic
down to the ptr so that all returnes will do size checks. A `into_raw_ptr_unchecked`
could be added in the future if a truely unchecked version is needed.

This change additionally removes the `try_` prefixes as the panic versions no longer
exist, and this makes the interfaces more consistent.