patina_internal_cpu: Always pop exception code when returning from interrupt
For x64, the current iplementation incorrectly tries to preserve the exception
code on the stack for interrupt vectors that originally pushed the code. The
SDM details in volume 3 section 6.13 that "Note that the error code is not popped
when the IRET instruction is executed to return from an exception handler, so the
handler must remove the error code before executing a return"

This changes the hint to be simply for alignment and always pops the exception code
whether it is real or not.