Add MaxAddress capability to memory manager (#688)
## Description
Add the ability to allocate with a maximum address for the memory
manager.

The immediate use case of this is to unblock ACPI component development,
although it may be useful for other memory-restricted operations.

- [X] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [X] Includes tests?
- [ ] Includes documentation?

## How This Was Tested
Added a new allocation as part of `memory_manager_allocations_test`.
Allocates and boots with no issues.

## Integration Instructions
N/A.

---------

Co-authored-by: Sherry Fan <sherryfan@microsoft.com>