systemtables: Checksum on specific protocol install (#690)
## Description

Adds a core component that registers an event that will re-checksum
the system tables when specific protocols (that would normally replace
functions in the system tables) are installed.

This fixes two SCT failures that were caused by the boot services
table and runtime services table having invalid checksums as we were
only performing a checksum on the tables before patina components or
UEFI spec compliant components were executed.
- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

N/A