patina_internal_collections: Update function signature (#711)
## Description

Updates the function signature of `get_mut` in both `bst.rs` and
`rbt.rs` to take a mutable reference to self. This change aligns with
Rust's safety principles, ensuring that mutable access to the data
structure is properly managed.

This should be considered a bug fix, as the previous signature was
invalid and could cause UB. No internal usage of `get_mut` being called
with an immutable reference was found, however.

closes #704

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI passes. Platform compiles. While this was a bug that clippy caught,
but it was not being actively misused.

## Integration Instructions

N/A