dxe_core: Switch TopDown to use max address for PI spec
The PI specified allocation type for TopDown specifies a max address and not a min address for both
memory and IO in section 7 as "EfiGcdAllocateMaxAddressSearchTopDown". The correct definition was used, but
converted into an inverse type. This commit resolves this inconsistency and converts all top down
searches to use a max instead of a min.