RFC Amendment: Add feature branch merge criteria to release process (#748)
## Description

Resolves #743 

Adds a section that describes the criteria for work on a feature branch
to be merged to the main branch.

The goal is to provide clarity so feature owners can plan how to
organize their work to support the requirements and the process of
adding new features is consistent over time for Patina consumers.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

- `cargo make all`

## Integration Instructions

- N/A

---------

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>