patina_debugger: restart the debugger if GdbStub crashes (#754)
The patina debugger will crash very often due to the internal GdbStub
state machine hitting an error. This is designed such that errors are
not recoverable. This commit makes it such that the debugger will be
restarted and the GdbStub re-initialized when this happens. This should
increase stability by avoiding debugger crashes.

When this restart occurs, any register state altered during the current
break would be lost. additionally, the IO queue will be cleared and the
break packet will be resent. It is unclear at this point if this
behavior will be problematic, so further refinement may be needed.