patina_dxe_core: Combine Core and UEFI dispatch (#762)
## Description

This commit combines the core component dispatcher and the UEFI driver
dispatcher into a single dispatch loop. This loop is broken into two
phases. The first phase performs a single pass of core component
dispatch while the second phase performs a single pass of UEFI driver
dispatch. Should either phase result in at least a single dispatch, the
loop will continue. This loop will only exit when a full iteration of
both phases results in no dispatches.

This functionality is necessary to produces services that are rust-ified
wrappers around C-functionality (such as protocols). While this is not a
suggested way of producing services, it is an acknowledged stop-gap
until pure rust implementations can be provided. This is specifically
necessary for #592, which requires certain protocols to exist before a
service can be produced (such as Console support, variable services,
etc).

Note: There is an intent to move the core UEFI dispatcher into a
multi-dispatchable component, however RFC work needs to be performed to
come up with an adequate solution for multi-dispatchable components, and
this is needed now to support the aforementioned RFC.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Verified boot to UEFI shell on virtual platform.

## Integration Instructions

N/A