GCD: Follow edk2 Attribute Behavior for Compat
edk2 has a behavior where if SetMemorySpaceAttributes() is
called with only virtual attributes set, then the page table
is not updated (as 0 is a valid value in the page table attrs: RWX).

Patina needs to follow this behavior to support existing drivers
that expect to be able to set a virtual attribute without applying
RWX to the range.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>