Increase stack space for AARCH64 synchronous exceptions. (#777)
## Description

The current implementation for the aarch64 exception handler reserves 8K
of stack for exceptions where the stack pointer may have been corrupted
at the time of exception. In practice (especially when optimizations are
disabled) this is not enough for all use cases, in particular when the
debugger is active.

This PR increases the space for the alternate stack to 64K. 

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Fixes debugger behavior on aarch64 platform used for test.

## Integration Instructions

N/A