docs/rust_tools: Add Patina tool overview (#792)
## Description

Adds a document to the Patina background section to explain the various
tools used in Patina and why they are used. To help readers better
relate these tools to what is commonly used in C firmware development,
some comparisons are made throughout the document.

The goal of this document is to explain key tools used in Patina, the
unifying interface behind the commands using cargo, and how these align
under a single Patina tool philosophy.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

- `cargo make all`
- Local markdown lint check

## Integration Instructions

- N/A

---------

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>