Handle potentially unaligned pointers from FFI (#786)
## Description

The UEFI spec does not clearly require that pointers passed to UEFI spec
APIs (like AllocatePages) must be aligned.
This pull request changes various FFI routines to assume that raw
pointers passed to them may not be aligned.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Unit tests pass. On systems passing unaligned pointers to core routines,
no longer observe crashes. Note: this doesn't constitute broad coverage
of these changes; this was only observed with a subset of the routines
in the real world.

## Integration Instructions

N/A