patina_dxe_core: Switch to use external RuntimeDxe implementation (#769)
The current core implements SetVirtualAddressMap within the core itself
and does not properly fill the runtime protocol. This means that when
the OS calls back into SetVirtualAddressMap, they are invoking code that
is executing in boot services memory and will often leverage boot
service memory allocated structures. This can cause problems as there is
no requirement that this memory is still valid and mapped at the point
where SetVirtualAddressMap was called.

This commit changes the runtime module to instead of exposing runtime
functions, to populate the runtime protocol produced by the RuntimeDxe
module to provide the event, and image data are available at runtime.
The expectation is now that there will be a separate module that will
implement the core runtime routines. For now, this will mean that
consumers will need to add the traditional RuntimeDxe driver or similar
to handle this logic.

This commit intentionally does not populate the memory map fields of the
runtime protocol as they are currently unused. This may change in the
future if other implementations are introduced.