Return Handles in Creation order; adjust verbosity for protocol install/uninstall messages. [Rebase & FF] (#797)
## Description

This PR makes two commits to change the protocol DB implementation:

Handle Database behavior updates
* Add support for tracking creation time for handles so that they are
returned in creation order for query functions like `LocateHandleBuffer`
* (minor) `registered_protocols` updated to return only unique protocols
(no duplicates)

Handle Database message verbosity
* (minor) Adjust verbosity of install/uninstall protocol interface
messages to align with traditional verbosity in EDK2 core.

Returning handles in creation order is not a UEFI spec requirement, but
returning them in creation order is behavior that a number of subsystems
(even within patina) rely on for reasonable behavior. An example within
patina is if multiple FVs are defined via HOBs at the start of DXE, they
should be dispatched in the order that they are reported to the core.
Without creation ordering on the handle query for new FVs, the
dispatcher will discover and process the FVs in arbitrary order.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Passes unit tests, executes on hardware platform without observed
issues.

## Integration Instructions

N/A