docs: Refactor platform integration docs (#830)
## Description

This is part 3 of 3 of a recent set of changes to documentation to
provide a broader perspective on Patina practices.

The other two for reference:

1. https://github.com/OpenDevicePartnership/patina/pull/757
2. https://github.com/OpenDevicePartnership/patina/pull/792

---

This PR focuses on platform integration. Originally, this was going to
add a new document to the
[`background`](https://github.com/OpenDevicePartnership/patina/tree/main/docs/src/background)
directory, but quite a bit of documentation for platform integration
already existed outside of the directory. Therefore, this task became
bringing focus to that documentation, updating it and remocing stale
concepts/content, and elaborating on some details to provide further
context. In particular, how platform integration is impacted today in an
EDK II workspace integrating the Patina DXE Core.

---

Restructures and updates platform integration content to remove stale
content and focus platform integration information. Platform details are
consolidated into a single document.

Removes:

- `concepts.md` - Covered concepts like LibraryClasses, PCD, and other
concepts that are not necessary to cover for Patina platform integration
some minor references to these concepts are included an "EDK II" of the
single document for those coming from EDK II.
- `integrate/compile_external.md` - Was a compilation guide for building
Patina in its own repo. This is the only supported model now. A separate
document is not needed.
- `integrate/compile_local.md` - Removed as this approach is now
deprecated.
- `integrate/platform_external.md` - Any relevant content was moved into
the unified document.

In `integrate/dxe_core.md`:

- Unified platform integration workflow from workspace setup to firmware
integration
- Expanded coverage of:
  - Dependency selection and minimal entry details
  - Core initialization and component registration procedures
  - Logging and debugging facility configuration
  - Platform-specific service and component integration guidance
  - Build and feature options with UEFI target compilation details
- EDK II integration process for replacing C DXE Core with Patina binary

This is meant to make it easier to find platform integration in a less
fragmented way, so it is easier to maintain and read.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

- `cargo make all`
- Markdown lint

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>