Update mdbook landing page to be clear and concise (#833)
## Description

The landing page is updated to provide a welcome page and minimal
information regarding what "Patina" is, and it's goals. It then links
across other pages in this book to provide more detailed information on
the different topics this book covers.

It also relocates the RFC Lifecycle and rust language information to
other parts of the book, as to not clutter the landing page. Some
additional information was added to the RFC Lifecycle page - regarding
when to use, and when not to use the RFC process.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

CI, review of the mdbook

## Integration Instructions

N/A