Remove "Configuration in Code" Best practices page (#836)
## Description

Remove the "Configuration in Code" best practices page as it is outdated
with the new (since written) configuration system for patina components.
The page being removed refers to what is now considered "private
configuration" (configuration specific to a component instance).

Documentation regarding public (shared) configuration and private
configuration is located in the Patina Component Model.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

CI

## Integration Instructions

N/A