Signal events in reverse order of creation (#838)
## Description

Presently Patina signals events in order of creation (assuming that they
are signaled at the same time with the same notify TPL). However, the
traditional EDK2 core signals events in reverse order of creation (i.e.
the most recently created events are signaled before older events). This
PR changes the signaling order to match EDK2 behavior.

The UEFI spec does not specify what order events should be signaled if
they are signaled at the same time and have the same notify TPL, but
this aligns Patina with existing implementations for maximum
compatibility.

This PR also adds a test to verify the ordering. 

This PR also contains a minor refactor of the `Ord::cmp` implementation
for `TaggedEventNotification`. This does not functionally change the
comparison order, but cleans up the style to avoid redundant
comparisons.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Existing and new unit tests pass, verified on platform hardware that
ordering of platform event signaling now matches EDK2 core.

## Integration Instructions

N/A