patina_dxe_core/pecoff: Verify PointerToRawData is resolved properly (#840)
## Description

Verifies that the behavior in the following goblin fix is followed:


https://github.com/m4b/goblin/commit/04d69ad5713d8a21aa295ef4ea047524a7edd615

This support is available in goblin
[0.10.2](https://crates.io/crates/goblin/0.10.2).

This verifies that the filename in the debug section is found using an
image built with 512 byte file alignment and one built with 32 byte file
alignment. Both have 4k-section alignment. This verifies that binaries
with very small file alignments have PointerToRawData calculated to the
correct position in the image.

The `goblin` version in Cargo.toml is set from `0.10` to `~0.10.2`. This
is to use versions >=0.10.2 and <0.11.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- `cargo make all`

## Integration Instructions

- N/A - New unit tests

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>