Do not set enable_patina_tests by default. (#885)
## Description

When it comes to compilation and feature flags for crates, cargo
attempts to find a version and feature configuration that satisfies all
usage of a single crate. When even a single one of our crates specify
that `enable_patina_tests` should be enabled, that ultimately enables
the feature for the platform, even if they did not request that it be
turned on.

To prevent this this feature from always being enabled, even when the
platform does not request it, we ensure that `enable_patina_tests` is
not enabled by default in any crate. This allows the platform developer
to toggle the feature in the patina crate and actually have it cascade
properly.

This commit fixes a bug that makes `enable_patina_tests` always enabled,
even when the platform does not add the feature flag to their `patina`
crate.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated that enabling the feature flag `enable_patina_tests` in the
platform repository (or disabling by not specifying `features =
["enable_patina_tests"]`) now correctly enables or disables the feature
flag. Previously the feature flag was always enabled, regardless of what
the platform said, because a dependent crate had it enabled, and the
cargo dependency resolver enabled it everywhere.

## Integration Instructions

N/A