patina_debugger: Fix OS lock check for aarch64 (#930)
When the OS lock register is set, certain debug features like single
stepping will not be allowed. This has been observed on aarch64 systems
where the OS lock is set initially. This fixes a typo on the initial
implementation that checked the wrong bit in the system register that
shows the status of the OS lock.