Change the image handle for the EFI System Table Pointer structure to EfiRuntimeServicesData. (#931)
## Description

This PR changes the image handle for the allocation for the EFI System
Table Pointer structure to be EfiBootServicesData; the prior use of the
DXE_CORE_HANDLE caused the region to appear as free in the memory map.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Confirmed windows boot with the change, verified memory map by
inspection.

## Integration Instructions

N/A