Update FromHob derive macro to utilize zerocopy (#943)
## Description

This commit updates the FromHob derive macro to rely on zerocopy to do
its conversion from a bytestream to a type. This requires that the
developer also derive implement zerocopy::FromBytes on the type. A
compile time error occurs if that type does not implement FromBytes.

**Please note:** there is no risk of trait version mismatches like we
have seen in the past as the entire usage of the trait (implementation
and consumption) is contained within this single file.

Additionally, this removes all `unsafe` code from the `FromHob` macro
expanded code.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [x] Includes tests?
- [x] Includes documentation?

## How This Was Tested

CI

## Integration Instructions

Users that currently use the `FromHob` derive macro must update their
interface as follows:

```rust
// OLD
#[derive(FromHob, Clone, Copy)]
#[hob = "guid"]
struct MyHob;

// NEW
#[derive(FromHob, zerocopy_derive::FromBytes)]
#[hob = "guid"]
struct MyHob;
```

Additionally, If the hob expects a `bool`, developers should switch to a
`u8` and evaluate the u8 to a boolean however they expect (Some expect
anything non-zero to be a boolean true, others expect `u8::MAX` to be a
boolean true, etc.