patina_debugger: Move TLB Flush to Monitor Command
Currently, the patina_debugger always issues a TLB flush
when exiting the debugger's active session. This can be
distracting when debugging issues with incorrect TLB entries
as using the debugger can make the problem go away. The
debugger does not need to do a TLB flush, when it edits the
page tables, the paging crate does all the required TLB
operations.

The reason the TLB flush was added to the debugger was to
handle the case where a developer is manually editing page
tables via the debugger client. This is a rare case, but does
require a TLB flush.

To handle this case in a less intrusive way, and to support
other uses cases where flushing the TLB is nice, this commit
drops always flushing the TLB from the debugger and instead
adds a monitor command to flush the TLB.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>